+++ src/features/flows/nRF9161/index.ts
@@ -43,13 +43,13 @@ const programConfig = [
                     type: 'program-modem-firmware',
                     firmware: {
                         core: 'Modem',
-                        file: 'mfw_nrf91x1_2.0.3.zip',
+                        file: 'mfw_nrf91x1_2.0.4.zip',
                         link: {
-                            label: 'Firmware v2.0.3',
-                            href: 'https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/sip/nrf91x1-sip/nrf91x1-lte-modem-firmware/release-notes/mfw_nrf91x1_2.0.3_release_notes.txt',
+                            label: 'Firmware v2.0.4',
+                            href: 'https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/sip/nrf91x1-sip/nrf91x1-lte-modem-firmware/release-notes/mfw_nrf91x1_2.0.4_release_notes.txt',
                         },
                     },
-                    version: '2.0.3',
+                    version: '2.0.4',
                     vComIndex: 0,
                 },
                 {
@@ -92,13 +92,13 @@ const programConfig = [
                     type: 'program-modem-firmware',
                     firmware: {
                         core: 'Modem',
-                        file: 'mfw_nrf91x1_2.0.3.zip',
+                        file: 'mfw_nrf91x1_2.0.4.zip',
                         link: {
-                            label: 'Firmware v2.0.3',
-                            href: 'https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/sip/nrf91x1-sip/nrf91x1-lte-modem-firmware/release-notes/mfw_nrf91x1_2.0.3_release_notes.txt',
+                            label: 'Firmware v2.0.4',
+                            href: 'https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/sip/nrf91x1-sip/nrf91x1-lte-modem-firmware/release-notes/mfw_nrf91x1_2.0.4_release_notes.txt',
                         },
                     },
-                    version: '2.0.3',
+                    version: '2.0.4',
                     vComIndex: 0,
                 },
                 {
@@ -141,13 +141,13 @@ const programConfig = [
                     type: 'program-modem-firmware',
                     firmware: {
                         core: 'Modem',
-                        file: 'mfw_nrf91x1_2.0.3.zip',
+                        file: 'mfw_nrf91x1_2.0.4.zip',
                         link: {
-                            label: 'Firmware v2.0.3',
-                            href: 'https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/sip/nrf91x1-sip/nrf91x1-lte-modem-firmware/release-notes/mfw_nrf91x1_2.0.3_release_notes.txt',
+                            label: 'Firmware v2.0.4',
+                            href: 'https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/sip/nrf91x1-sip/nrf91x1-lte-modem-firmware/release-notes/mfw_nrf91x1_2.0.4_release_notes.txt',
                         },
                     },
-                    version: '2.0.3',
+                    version: '2.0.4',
                     vComIndex: 0,
                 },
                 {
@@ -191,13 +191,13 @@ const programConfig = [
                     type: 'program-modem-firmware',
                     firmware: {
                         core: 'Modem',
-                        file: 'mfw_nrf91x1_2.0.3.zip',
+                        file: 'mfw_nrf91x1_2.0.4.zip',
                         link: {
-                            label: 'Firmware v2.0.3',
-                            href: 'https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/sip/nrf91x1-sip/nrf91x1-lte-modem-firmware/release-notes/mfw_nrf91x1_2.0.3_release_notes.txt',
+                            label: 'Firmware v2.0.4',
+                            href: 'https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/sip/nrf91x1-sip/nrf91x1-lte-modem-firmware/release-notes/mfw_nrf91x1_2.0.4_release_notes.txt',
                         },
                     },
-                    version: '2.0.3',
+                    version: '2.0.4',
                     vComIndex: 0,
                 },
                 {